Adapt to the new structure
