--- project-api/src/api/Jalasoft.TeamUp.Projects.API/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "Jalasoft.TeamUp.Projects.API": {
-      "commandName": "Project",
-      "commandLineArgs": "host start --pause--on-error",
-      "nativeDebugging": false,
-      "jsWebView2Debugging": false
-    }
-  }
-}
