+++ src/core/Jalasoft.TeamUp.Projects.Core.Interfaces/IProjectsService.cs
