+++ src/core/Jalasoft.TeamUp.Projects.Core/ProjectsService.cs
