+++ src/dal/Jalasoft.TeamUp.Projects.DAL.Interfaces/IProjectsRepository.cs
