+++ src/dal/Jalasoft.TeamUp.Projects.DAL/Jalasoft.TeamUp.Projects.DAL.csproj
@@ -13,6 +13,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <ProjectReference Include="..\..\models\Jalasoft.TeamUp.Projects.Models\Jalasoft.TeamUp.Projects.Models.csproj" />
 	  <ProjectReference Include="..\Jalasoft.TeamUp.Projects.DAL.Interfaces\Jalasoft.TeamUp.Projects.DAL.Interfaces.csproj" />
 	</ItemGroup>
