+++ src/models/Jalasoft.TeamUp.Projects.Models/Contact.cs
@@ -1,0 +1,11 @@
+﻿namespace Jalasoft.TeamUp.Projects.Models
+{
+    using System;
+
+    public class Contact
+    {
+        public string Name { get; set; }
+
+        public Guid IdResume { get; set; }
+    }
+}
