+++ src/models/Jalasoft.TeamUp.Projects.Models/Project.cs
@@ -12,6 +12,16 @@
 
         public string Description { get; set; }
 
-        public string Contact { get; set; }
+        public Contact Contact { get; set; }
+
+        public bool State { get; set; }
+
+        public string TextInvitation { get; set; }
+
+        public string Logo { get; set; }
+
+        public List<Contact> MemberList { get; set; }
+
+        public DateTime CreationDate { get; set; }
     }
 }
