--- tests/api/Jalasoft.TeamUp.Projects.API.Tests/UnitTest1.cs
@@ -1,13 +1,0 @@
-namespace Jalasoft.TeamUp.Projects.API.Tests
-{
-    using System;
-    using Xunit;
-
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-        }
-    }
-}
